requirements.txt
Flask
pillow
numpy
tensorflow
